rename package (fte: flat torus embedding)
