+++ src/test/scala/fte/data/Edge.java
@@ -1,4 +1,4 @@
-package vmi.graph.data;
+package fte.data;
 
 
 public class Edge implements Cloneable {
