+++ src/test/scala/fte/data/Face.java
@@ -1,4 +1,4 @@
-package vmi.graph.data;
+package fte.data;
 
 import java.util.LinkedList;
